roadsense-core/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
python-multipart
pydantic
